- Implemented an Input binding manager (press F1, in Input section) to assign custom keys and gamepad/joystick axis and buttons - Added a `counter` node in computer - Improved Launcher: Downloading and auto-updating flow and performance has been improved greatly - The game can now be played in Offline mode as a guest player (if you have an existing world) - Solar panels now take into account all radiation from the sun(s) and its relative angle (but no occlusions yet) - Bugfix: screenshots were skewed when resolution width was not a multiple of 8 - Fixed re-entry effect visibility on other (non-parent) vehicles - Fixed avatar collision detection with components of other (non-parent) vehicles
This is the last update on the techdemo branch of Archean.
We are now working strickly on the Alpha release branch which will NOT be compatible with worlds created using the techdemo branch.
Those who have bought the game during the techdemo WILL have full access to the Alpha branch once we release it in a few weeks.